examples/python-example/requirements.txt
pyaudio
python-socketio[client]
websocket-client
